+++ lib/processor/cache.go
@@ -55,7 +55,7 @@ pipeline:
         value: "storeme"
     - bloblang: root = if errored() { deleted() }
 
-resource_caches:
+cache_resources:
   - label: foocache
     redis:
       url: tcp://TODO:6379
@@ -77,7 +77,7 @@ pipeline:
               key: '${! json("message.document_id") }'
         result_map: 'root.message.document = this'
 
-resource_caches:
+cache_resources:
   - label: foocache
     memcached:
       addresses: [ "TODO:11211" ]