+++ lib/test/integration/cache/redis_cluster_test.go
@@ -82,7 +82,7 @@ var _ = registerIntegrationTest("redis_cluster", func(t *testing.T) {
 	}))
 
 	template := `
-resource_caches:
+cache_resources:
   - label: testcache
     redis:
       url: $VAR1